core: regenerate entities with latest scheme changes
