--- library/Ivoz/Cgr/Domain/Model/TpDerivedCharge/TpDerivedChargeDto.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Ivoz\Cgr\Domain\Model\TpDerivedCharge;
-
-
-class TpDerivedChargeDto extends TpDerivedChargeDtoAbstract
-{
-
-}
-
-
