--- library/Ivoz/Cgr/Domain/Model/TpDerivedCharge/TpDerivedChargeRepository.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Ivoz\Cgr\Domain\Model\TpDerivedCharge;
-
-use Doctrine\Common\Persistence\ObjectRepository;
-use Doctrine\Common\Collections\Selectable;
-
-interface TpDerivedChargeRepository extends  ObjectRepository, Selectable
-{
-
-}
-
-
