+++ library/Ivoz/Cgr/Domain/Model/TpDestinationRate/TpDestinationRateDtoAbstract.php
@@ -63,49 +63,9 @@ abstract class TpDestinationRateDtoAbstract implements DataTransferObjectInterfa
     private $id;
 
     /**
-     * @var string
-     */
-    private $destinationPrefix;
-
-    /**
-     * @var string
-     */
-    private $destinationPrefixName;
-
-    /**
-     * @var string
-     */
-    private $rateCost;
-
-    /**
-     * @var string
-     */
-    private $rateConnectFee;
-
-    /**
-     * @var string
-     */
-    private $rateRateIncrement;
-
-    /**
-     * @var string
-     */
-    private $rateGroupIntervalStart = '0s';
-
-    /**
      * @var \Ivoz\Provider\Domain\Model\DestinationRate\DestinationRateDto | null
      */
     private $destinationRate;
-
-    /**
-     * @var \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto | null
-     */
-    private $tpDestination;
-
-    /**
-     * @var \Ivoz\Cgr\Domain\Model\TpRate\TpRateDto | null
-     */
-    private $tpRate;
 
 
     use DtoNormalizer;
@@ -135,11 +95,7 @@ abstract class TpDestinationRateDtoAbstract implements DataTransferObjectInterfa
             'maxCostStrategy' => 'maxCostStrategy',
             'createdAt' => 'createdAt',
             'id' => 'id',
-            'destination' => ['prefix','prefixName'],
-            'rate' => ['cost','connectFee','rateIncrement','groupIntervalStart'],
-            'destinationRateId' => 'destinationRate',
-            'tpDestinationId' => 'tpDestination',
-            'tpRateId' => 'tpRate'
+            'destinationRateId' => 'destinationRate'
         ];
     }
 
@@ -159,19 +115,7 @@ abstract class TpDestinationRateDtoAbstract implements DataTransferObjectInterfa
             'maxCostStrategy' => $this->getMaxCostStrategy(),
             'createdAt' => $this->getCreatedAt(),
             'id' => $this->getId(),
-            'destination' => [
-                'prefix' => $this->getDestinationPrefix(),
-                'prefixName' => $this->getDestinationPrefixName()
-            ],
-            'rate' => [
-                'cost' => $this->getRateCost(),
-                'connectFee' => $this->getRateConnectFee(),
-                'rateIncrement' => $this->getRateRateIncrement(),
-                'groupIntervalStart' => $this->getRateGroupIntervalStart()
-            ],
-            'destinationRate' => $this->getDestinationRate(),
-            'tpDestination' => $this->getTpDestination(),
-            'tpRate' => $this->getTpRate()
+            'destinationRate' => $this->getDestinationRate()
         ];
     }
 
@@ -181,8 +125,6 @@ abstract class TpDestinationRateDtoAbstract implements DataTransferObjectInterfa
     public function transformForeignKeys(ForeignKeyTransformerInterface $transformer)
     {
         $this->destinationRate = $transformer->transform('Ivoz\\Provider\\Domain\\Model\\DestinationRate\\DestinationRate', $this->getDestinationRateId());
-        $this->tpDestination = $transformer->transform('Ivoz\\Cgr\\Domain\\Model\\TpDestination\\TpDestination', $this->getTpDestinationId());
-        $this->tpRate = $transformer->transform('Ivoz\\Cgr\\Domain\\Model\\TpRate\\TpRate', $this->getTpRateId());
     }
 
     /**
@@ -394,126 +336,6 @@ abstract class TpDestinationRateDtoAbstract implements DataTransferObjectInterfa
     }
 
     /**
-     * @param string $destinationPrefix
-     *
-     * @return static
-     */
-    public function setDestinationPrefix($destinationPrefix = null)
-    {
-        $this->destinationPrefix = $destinationPrefix;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDestinationPrefix()
-    {
-        return $this->destinationPrefix;
-    }
-
-    /**
-     * @param string $destinationPrefixName
-     *
-     * @return static
-     */
-    public function setDestinationPrefixName($destinationPrefixName = null)
-    {
-        $this->destinationPrefixName = $destinationPrefixName;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDestinationPrefixName()
-    {
-        return $this->destinationPrefixName;
-    }
-
-    /**
-     * @param string $rateCost
-     *
-     * @return static
-     */
-    public function setRateCost($rateCost = null)
-    {
-        $this->rateCost = $rateCost;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getRateCost()
-    {
-        return $this->rateCost;
-    }
-
-    /**
-     * @param string $rateConnectFee
-     *
-     * @return static
-     */
-    public function setRateConnectFee($rateConnectFee = null)
-    {
-        $this->rateConnectFee = $rateConnectFee;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getRateConnectFee()
-    {
-        return $this->rateConnectFee;
-    }
-
-    /**
-     * @param string $rateRateIncrement
-     *
-     * @return static
-     */
-    public function setRateRateIncrement($rateRateIncrement = null)
-    {
-        $this->rateRateIncrement = $rateRateIncrement;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getRateRateIncrement()
-    {
-        return $this->rateRateIncrement;
-    }
-
-    /**
-     * @param string $rateGroupIntervalStart
-     *
-     * @return static
-     */
-    public function setRateGroupIntervalStart($rateGroupIntervalStart = null)
-    {
-        $this->rateGroupIntervalStart = $rateGroupIntervalStart;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getRateGroupIntervalStart()
-    {
-        return $this->rateGroupIntervalStart;
-    }
-
-    /**
      * @param \Ivoz\Provider\Domain\Model\DestinationRate\DestinationRateDto $destinationRate
      *
      * @return static
@@ -553,98 +375,6 @@ abstract class TpDestinationRateDtoAbstract implements DataTransferObjectInterfa
     public function getDestinationRateId()
     {
         if ($dto = $this->getDestinationRate()) {
-            return $dto->getId();
-        }
-
-        return null;
-    }
-
-    /**
-     * @param \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto $tpDestination
-     *
-     * @return static
-     */
-    public function setTpDestination(\Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto $tpDestination = null)
-    {
-        $this->tpDestination = $tpDestination;
-
-        return $this;
-    }
-
-    /**
-     * @return \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto
-     */
-    public function getTpDestination()
-    {
-        return $this->tpDestination;
-    }
-
-    /**
-     * @param integer $id | null
-     *
-     * @return static
-     */
-    public function setTpDestinationId($id)
-    {
-        $value = !is_null($id)
-            ? new \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto($id)
-            : null;
-
-        return $this->setTpDestination($value);
-    }
-
-    /**
-     * @return integer | null
-     */
-    public function getTpDestinationId()
-    {
-        if ($dto = $this->getTpDestination()) {
-            return $dto->getId();
-        }
-
-        return null;
-    }
-
-    /**
-     * @param \Ivoz\Cgr\Domain\Model\TpRate\TpRateDto $tpRate
-     *
-     * @return static
-     */
-    public function setTpRate(\Ivoz\Cgr\Domain\Model\TpRate\TpRateDto $tpRate = null)
-    {
-        $this->tpRate = $tpRate;
-
-        return $this;
-    }
-
-    /**
-     * @return \Ivoz\Cgr\Domain\Model\TpRate\TpRateDto
-     */
-    public function getTpRate()
-    {
-        return $this->tpRate;
-    }
-
-    /**
-     * @param integer $id | null
-     *
-     * @return static
-     */
-    public function setTpRateId($id)
-    {
-        $value = !is_null($id)
-            ? new \Ivoz\Cgr\Domain\Model\TpRate\TpRateDto($id)
-            : null;
-
-        return $this->setTpRate($value);
-    }
-
-    /**
-     * @return integer | null
-     */
-    public function getTpRateId()
-    {
-        if ($dto = $this->getTpRate()) {
             return $dto->getId();
         }
