+++ library/Ivoz/Kam/Domain/Model/TrunksCdr/TrunksCdrAbstract.php
@@ -115,16 +115,6 @@ abstract class TrunksCdrAbstract
      */
     protected $carrier;
 
-    /**
-     * @var \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationInterface
-     */
-    protected $tpDestination;
-
-    /**
-     * @var \Ivoz\Provider\Domain\Model\DestinationRate\DestinationRateInterface
-     */
-    protected $destinationRate;
-
 
     use ChangelogTrait;
 
@@ -224,8 +214,6 @@ abstract class TrunksCdrAbstract
             ->setBrand($dto->getBrand())
             ->setCompany($dto->getCompany())
             ->setCarrier($dto->getCarrier())
-            ->setTpDestination($dto->getTpDestination())
-            ->setDestinationRate($dto->getDestinationRate())
         ;
 
         $self->sanitizeValues();
@@ -265,9 +253,7 @@ abstract class TrunksCdrAbstract
             ->setInvoice($dto->getInvoice())
             ->setBrand($dto->getBrand())
             ->setCompany($dto->getCompany())
-            ->setCarrier($dto->getCarrier())
-            ->setTpDestination($dto->getTpDestination())
-            ->setDestinationRate($dto->getDestinationRate());
+            ->setCarrier($dto->getCarrier());
 
 
 
@@ -301,9 +287,7 @@ abstract class TrunksCdrAbstract
             ->setInvoice(\Ivoz\Provider\Domain\Model\Invoice\Invoice::entityToDto(self::getInvoice(), $depth))
             ->setBrand(\Ivoz\Provider\Domain\Model\Brand\Brand::entityToDto(self::getBrand(), $depth))
             ->setCompany(\Ivoz\Provider\Domain\Model\Company\Company::entityToDto(self::getCompany(), $depth))
-            ->setCarrier(\Ivoz\Provider\Domain\Model\Carrier\Carrier::entityToDto(self::getCarrier(), $depth))
-            ->setTpDestination(\Ivoz\Cgr\Domain\Model\TpDestination\TpDestination::entityToDto(self::getTpDestination(), $depth))
-            ->setDestinationRate(\Ivoz\Provider\Domain\Model\DestinationRate\DestinationRate::entityToDto(self::getDestinationRate(), $depth));
+            ->setCarrier(\Ivoz\Provider\Domain\Model\Carrier\Carrier::entityToDto(self::getCarrier(), $depth));
     }
 
     /**
@@ -331,9 +315,7 @@ abstract class TrunksCdrAbstract
             'invoiceId' => self::getInvoice() ? self::getInvoice()->getId() : null,
             'brandId' => self::getBrand() ? self::getBrand()->getId() : null,
             'companyId' => self::getCompany() ? self::getCompany()->getId() : null,
-            'carrierId' => self::getCarrier() ? self::getCarrier()->getId() : null,
-            'tpDestinationId' => self::getTpDestination() ? self::getTpDestination()->getId() : null,
-            'destinationRateId' => self::getDestinationRate() ? self::getDestinationRate()->getId() : null
+            'carrierId' => self::getCarrier() ? self::getCarrier()->getId() : null
         ];
     }
 
@@ -886,54 +868,6 @@ abstract class TrunksCdrAbstract
     public function getCarrier()
     {
         return $this->carrier;
-    }
-
-    /**
-     * Set tpDestination
-     *
-     * @param \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationInterface $tpDestination
-     *
-     * @return self
-     */
-    public function setTpDestination(\Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationInterface $tpDestination = null)
-    {
-        $this->tpDestination = $tpDestination;
-
-        return $this;
-    }
-
-    /**
-     * Get tpDestination
-     *
-     * @return \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationInterface
-     */
-    public function getTpDestination()
-    {
-        return $this->tpDestination;
-    }
-
-    /**
-     * Set destinationRate
-     *
-     * @param \Ivoz\Provider\Domain\Model\DestinationRate\DestinationRateInterface $destinationRate
-     *
-     * @return self
-     */
-    public function setDestinationRate(\Ivoz\Provider\Domain\Model\DestinationRate\DestinationRateInterface $destinationRate = null)
-    {
-        $this->destinationRate = $destinationRate;
-
-        return $this;
-    }
-
-    /**
-     * Get destinationRate
-     *
-     * @return \Ivoz\Provider\Domain\Model\DestinationRate\DestinationRateInterface
-     */
-    public function getDestinationRate()
-    {
-        return $this->destinationRate;
     }
