+++ library/Ivoz/Provider/Domain/Model/Destination/DestinationDto.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Ivoz\Provider\Domain\Model\Destination;
+
+
+class DestinationDto extends DestinationDtoAbstract
+{
+
+}
+
+
