+++ library/Ivoz/Provider/Domain/Model/Destination/DestinationDtoAbstract.php
@@ -1,0 +1,318 @@
+<?php
+
+namespace Ivoz\Provider\Domain\Model\Destination;
+
+use Ivoz\Core\Application\DataTransferObjectInterface;
+use Ivoz\Core\Application\ForeignKeyTransformerInterface;
+use Ivoz\Core\Application\CollectionTransformerInterface;
+use Ivoz\Core\Application\Model\DtoNormalizer;
+
+/**
+ * @codeCoverageIgnore
+ */
+abstract class DestinationDtoAbstract implements DataTransferObjectInterface
+{
+    /**
+     * @var string
+     */
+    private $prefix;
+
+    /**
+     * @var integer
+     */
+    private $id;
+
+    /**
+     * @var string
+     */
+    private $nameEn;
+
+    /**
+     * @var string
+     */
+    private $nameEs;
+
+    /**
+     * @var \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto | null
+     */
+    private $tpDestination;
+
+    /**
+     * @var \Ivoz\Provider\Domain\Model\Brand\BrandDto | null
+     */
+    private $brand;
+
+    /**
+     * @var \Ivoz\Provider\Domain\Model\DestinationRate\DestinationRateDto[] | null
+     */
+    private $destinationRates = null;
+
+
+    use DtoNormalizer;
+
+    public function __construct($id = null)
+    {
+        $this->setId($id);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public static function getPropertyMap(string $context = '')
+    {
+        if ($context === self::CONTEXT_COLLECTION) {
+            return ['id' => 'id'];
+        }
+
+        return [
+            'prefix' => 'prefix',
+            'id' => 'id',
+            'name' => ['en','es'],
+            'tpDestinationId' => 'tpDestination',
+            'brandId' => 'brand'
+        ];
+    }
+
+    /**
+     * @return array
+     */
+    public function toArray($hideSensitiveData = false)
+    {
+        return [
+            'prefix' => $this->getPrefix(),
+            'id' => $this->getId(),
+            'name' => [
+                'en' => $this->getNameEn(),
+                'es' => $this->getNameEs()
+            ],
+            'tpDestination' => $this->getTpDestination(),
+            'brand' => $this->getBrand(),
+            'destinationRates' => $this->getDestinationRates()
+        ];
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function transformForeignKeys(ForeignKeyTransformerInterface $transformer)
+    {
+        $this->tpDestination = $transformer->transform('Ivoz\\Cgr\\Domain\\Model\\TpDestination\\TpDestination', $this->getTpDestinationId());
+        $this->brand = $transformer->transform('Ivoz\\Provider\\Domain\\Model\\Brand\\Brand', $this->getBrandId());
+        if (!is_null($this->destinationRates)) {
+            $items = $this->getDestinationRates();
+            $this->destinationRates = [];
+            foreach ($items as $item) {
+                $this->destinationRates[] = $transformer->transform(
+                    'Ivoz\\Provider\\Domain\\Model\\DestinationRate\\DestinationRate',
+                    $item->getId() ?? $item
+                );
+            }
+        }
+
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function transformCollections(CollectionTransformerInterface $transformer)
+    {
+        $this->destinationRates = $transformer->transform(
+            'Ivoz\\Provider\\Domain\\Model\\DestinationRate\\DestinationRate',
+            $this->destinationRates
+        );
+    }
+
+    /**
+     * @param string $prefix
+     *
+     * @return static
+     */
+    public function setPrefix($prefix = null)
+    {
+        $this->prefix = $prefix;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getPrefix()
+    {
+        return $this->prefix;
+    }
+
+    /**
+     * @param integer $id
+     *
+     * @return static
+     */
+    public function setId($id = null)
+    {
+        $this->id = $id;
+
+        return $this;
+    }
+
+    /**
+     * @return integer
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param string $nameEn
+     *
+     * @return static
+     */
+    public function setNameEn($nameEn = null)
+    {
+        $this->nameEn = $nameEn;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getNameEn()
+    {
+        return $this->nameEn;
+    }
+
+    /**
+     * @param string $nameEs
+     *
+     * @return static
+     */
+    public function setNameEs($nameEs = null)
+    {
+        $this->nameEs = $nameEs;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getNameEs()
+    {
+        return $this->nameEs;
+    }
+
+    /**
+     * @param \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto $tpDestination
+     *
+     * @return static
+     */
+    public function setTpDestination(\Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto $tpDestination = null)
+    {
+        $this->tpDestination = $tpDestination;
+
+        return $this;
+    }
+
+    /**
+     * @return \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto
+     */
+    public function getTpDestination()
+    {
+        return $this->tpDestination;
+    }
+
+    /**
+     * @param integer $id | null
+     *
+     * @return static
+     */
+    public function setTpDestinationId($id)
+    {
+        $value = !is_null($id)
+            ? new \Ivoz\Cgr\Domain\Model\TpDestination\TpDestinationDto($id)
+            : null;
+
+        return $this->setTpDestination($value);
+    }
+
+    /**
+     * @return integer | null
+     */
+    public function getTpDestinationId()
+    {
+        if ($dto = $this->getTpDestination()) {
+            return $dto->getId();
+        }
+
+        return null;
+    }
+
+    /**
+     * @param \Ivoz\Provider\Domain\Model\Brand\BrandDto $brand
+     *
+     * @return static
+     */
+    public function setBrand(\Ivoz\Provider\Domain\Model\Brand\BrandDto $brand = null)
+    {
+        $this->brand = $brand;
+
+        return $this;
+    }
+
+    /**
+     * @return \Ivoz\Provider\Domain\Model\Brand\BrandDto
+     */
+    public function getBrand()
+    {
+        return $this->brand;
+    }
+
+    /**
+     * @param integer $id | null
+     *
+     * @return static
+     */
+    public function setBrandId($id)
+    {
+        $value = !is_null($id)
+            ? new \Ivoz\Provider\Domain\Model\Brand\BrandDto($id)
+            : null;
+
+        return $this->setBrand($value);
+    }
+
+    /**
+     * @return integer | null
+     */
+    public function getBrandId()
+    {
+        if ($dto = $this->getBrand()) {
+            return $dto->getId();
+        }
+
+        return null;
+    }
+
+    /**
+     * @param array $destinationRates
+     *
+     * @return static
+     */
+    public function setDestinationRates($destinationRates = null)
+    {
+        $this->destinationRates = $destinationRates;
+
+        return $this;
+    }
+
+    /**
+     * @return array
+     */
+    public function getDestinationRates()
+    {
+        return $this->destinationRates;
+    }
+}
+
+
