+++ library/Ivoz/Provider/Domain/Model/DestinationRate/DestinationRateDto.php
@@ -5,32 +5,7 @@ namespace Ivoz\Provider\Domain\Model\DestinationRate;
 
 class DestinationRateDto extends DestinationRateDtoAbstract
 {
-    private $filePath;
 
-    public function getFileObjects()
-    {
-        return [
-            'File'
-        ];
-    }
-
-    /**
-     * @return self
-     */
-    public function setFilePath(string $path = null)
-    {
-        $this->filePath = $path;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getFilePath()
-    {
-        return $this->filePath;
-    }
 }
