+++ library/Ivoz/Provider/Domain/Model/DestinationRate/DestinationRateRepository.php
@@ -5,8 +5,9 @@ namespace Ivoz\Provider\Domain\Model\DestinationRate;
 use Doctrine\Common\Persistence\ObjectRepository;
 use Doctrine\Common\Collections\Selectable;
 
-interface DestinationRateRepository extends ObjectRepository, Selectable
+interface DestinationRateRepository extends  ObjectRepository, Selectable
 {
 
 }
+
