+++ library/Ivoz/Provider/Domain/Model/DestinationRate/DestinationRateTrait.php
@@ -3,9 +3,6 @@
 namespace Ivoz\Provider\Domain\Model\DestinationRate;
 
 use Ivoz\Core\Application\DataTransferObjectInterface;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\Common\Collections\Criteria;
 
 /**
  * DestinationRateTrait
@@ -18,11 +15,6 @@ trait DestinationRateTrait
      */
     protected $id;
 
-    /**
-     * @var Collection
-     */
-    protected $tpDestinationRates;
-
 
     /**
      * Constructor
@@ -30,7 +22,7 @@ trait DestinationRateTrait
     protected function __construct()
     {
         parent::__construct(...func_get_args());
-        $this->tpDestinationRates = new ArrayCollection();
+
     }
 
     /**
@@ -44,9 +36,7 @@ trait DestinationRateTrait
          * @var $dto DestinationRateDto
          */
         $self = parent::fromDto($dto);
-        if ($dto->getTpDestinationRates()) {
-            $self->replaceTpDestinationRates($dto->getTpDestinationRates());
-        }
+
         if ($dto->getId()) {
             $self->id = $dto->getId();
             $self->initChangelog();
@@ -65,9 +55,7 @@ trait DestinationRateTrait
          * @var $dto DestinationRateDto
          */
         parent::updateFromDto($dto);
-        if ($dto->getTpDestinationRates()) {
-            $this->replaceTpDestinationRates($dto->getTpDestinationRates());
-        }
+
         return $this;
     }
 
@@ -90,79 +78,6 @@ trait DestinationRateTrait
         return parent::__toArray() + [
             'id' => self::getId()
         ];
-    }
-
-
-    /**
-     * Add tpDestinationRate
-     *
-     * @param \Ivoz\Cgr\Domain\Model\TpDestinationRate\TpDestinationRateInterface $tpDestinationRate
-     *
-     * @return DestinationRateTrait
-     */
-    public function addTpDestinationRate(\Ivoz\Cgr\Domain\Model\TpDestinationRate\TpDestinationRateInterface $tpDestinationRate)
-    {
-        $this->tpDestinationRates->add($tpDestinationRate);
-
-        return $this;
-    }
-
-    /**
-     * Remove tpDestinationRate
-     *
-     * @param \Ivoz\Cgr\Domain\Model\TpDestinationRate\TpDestinationRateInterface $tpDestinationRate
-     */
-    public function removeTpDestinationRate(\Ivoz\Cgr\Domain\Model\TpDestinationRate\TpDestinationRateInterface $tpDestinationRate)
-    {
-        $this->tpDestinationRates->removeElement($tpDestinationRate);
-    }
-
-    /**
-     * Replace tpDestinationRates
-     *
-     * @param \Ivoz\Cgr\Domain\Model\TpDestinationRate\TpDestinationRateInterface[] $tpDestinationRates
-     * @return self
-     */
-    public function replaceTpDestinationRates(Collection $tpDestinationRates)
-    {
-        $updatedEntities = [];
-        $fallBackId = -1;
-        foreach ($tpDestinationRates as $entity) {
-            $index = $entity->getId() ? $entity->getId() : $fallBackId--;
-            $updatedEntities[$index] = $entity;
-            $entity->setDestinationRate($this);
-        }
-        $updatedEntityKeys = array_keys($updatedEntities);
-
-        foreach ($this->tpDestinationRates as $key => $entity) {
-            $identity = $entity->getId();
-            if (in_array($identity, $updatedEntityKeys)) {
-                $this->tpDestinationRates->set($key, $updatedEntities[$identity]);
-            } else {
-                $this->tpDestinationRates->remove($key);
-            }
-            unset($updatedEntities[$identity]);
-        }
-
-        foreach ($updatedEntities as $entity) {
-            $this->addTpDestinationRate($entity);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Get tpDestinationRates
-     *
-     * @return \Ivoz\Cgr\Domain\Model\TpDestinationRate\TpDestinationRateInterface[]
-     */
-    public function getTpDestinationRates(Criteria $criteria = null)
-    {
-        if (!is_null($criteria)) {
-            return $this->tpDestinationRates->matching($criteria)->toArray();
-        }
-
-        return $this->tpDestinationRates->toArray();
     }
