+++ library/Ivoz/Provider/Domain/Model/DestinationRateGroup/DestinationRateGroupAbstract.php
@@ -1,0 +1,369 @@
+<?php
+
+namespace Ivoz\Provider\Domain\Model\DestinationRateGroup;
+
+use Assert\Assertion;
+use Ivoz\Core\Application\DataTransferObjectInterface;
+use Ivoz\Core\Domain\Model\ChangelogTrait;
+use Ivoz\Core\Domain\Model\EntityInterface;
+
+/**
+ * DestinationRateGroupAbstract
+ * @codeCoverageIgnore
+ */
+abstract class DestinationRateGroupAbstract
+{
+    /**
+     * comment: enum:waiting|inProgress|imported|error
+     * @var string
+     */
+    protected $status;
+
+    /**
+     * @var Name
+     */
+    protected $name;
+
+    /**
+     * @var Description
+     */
+    protected $description;
+
+    /**
+     * @var File
+     */
+    protected $file;
+
+    /**
+     * @var \Ivoz\Provider\Domain\Model\Brand\BrandInterface
+     */
+    protected $brand;
+
+
+    use ChangelogTrait;
+
+    /**
+     * Constructor
+     */
+    protected function __construct(
+        Name $name,
+        Description $description,
+        File $file
+    ) {
+        $this->setName($name);
+        $this->setDescription($description);
+        $this->setFile($file);
+    }
+
+    abstract public function getId();
+
+    public function __toString()
+    {
+        return sprintf("%s#%s",
+            "DestinationRateGroup",
+            $this->getId()
+        );
+    }
+
+    /**
+     * @return void
+     * @throws \Exception
+     */
+    protected function sanitizeValues()
+    {
+    }
+
+    /**
+     * @param null $id
+     * @return DestinationRateGroupDto
+     */
+    public static function createDto($id = null)
+    {
+        return new DestinationRateGroupDto($id);
+    }
+
+    /**
+     * @param EntityInterface|null $entity
+     * @param int $depth
+     * @return DestinationRateGroupDto|null
+     */
+    public static function entityToDto(EntityInterface $entity = null, $depth = 0)
+    {
+        if (!$entity) {
+            return null;
+        }
+
+        Assertion::isInstanceOf($entity, DestinationRateGroupInterface::class);
+
+        if ($depth < 1) {
+            return static::createDto($entity->getId());
+        }
+
+        if ($entity instanceof \Doctrine\ORM\Proxy\Proxy && !$entity->__isInitialized()) {
+            return static::createDto($entity->getId());
+        }
+
+        return $entity->toDto($depth-1);
+    }
+
+    /**
+     * Factory method
+     * @param DataTransferObjectInterface $dto
+     * @return self
+     */
+    public static function fromDto(DataTransferObjectInterface $dto)
+    {
+        /**
+         * @var $dto DestinationRateGroupDto
+         */
+        Assertion::isInstanceOf($dto, DestinationRateGroupDto::class);
+
+        $name = new Name(
+            $dto->getNameEn(),
+            $dto->getNameEs()
+        );
+
+        $description = new Description(
+            $dto->getDescriptionEn(),
+            $dto->getDescriptionEs()
+        );
+
+        $file = new File(
+            $dto->getFileFileSize(),
+            $dto->getFileMimeType(),
+            $dto->getFileBaseName(),
+            $dto->getFileImporterArguments()
+        );
+
+        $self = new static(
+            $name,
+            $description,
+            $file
+        );
+
+        $self
+            ->setStatus($dto->getStatus())
+            ->setBrand($dto->getBrand())
+        ;
+
+        $self->sanitizeValues();
+        $self->initChangelog();
+
+        return $self;
+    }
+
+    /**
+     * @param DataTransferObjectInterface $dto
+     * @return self
+     */
+    public function updateFromDto(DataTransferObjectInterface $dto)
+    {
+        /**
+         * @var $dto DestinationRateGroupDto
+         */
+        Assertion::isInstanceOf($dto, DestinationRateGroupDto::class);
+
+        $name = new Name(
+            $dto->getNameEn(),
+            $dto->getNameEs()
+        );
+
+        $description = new Description(
+            $dto->getDescriptionEn(),
+            $dto->getDescriptionEs()
+        );
+
+        $file = new File(
+            $dto->getFileFileSize(),
+            $dto->getFileMimeType(),
+            $dto->getFileBaseName(),
+            $dto->getFileImporterArguments()
+        );
+
+        $this
+            ->setStatus($dto->getStatus())
+            ->setName($name)
+            ->setDescription($description)
+            ->setFile($file)
+            ->setBrand($dto->getBrand());
+
+
+
+        $this->sanitizeValues();
+        return $this;
+    }
+
+    /**
+     * @param int $depth
+     * @return DestinationRateGroupDto
+     */
+    public function toDto($depth = 0)
+    {
+        return self::createDto()
+            ->setStatus(self::getStatus())
+            ->setNameEn(self::getName()->getEn())
+            ->setNameEs(self::getName()->getEs())
+            ->setDescriptionEn(self::getDescription()->getEn())
+            ->setDescriptionEs(self::getDescription()->getEs())
+            ->setFileFileSize(self::getFile()->getFileSize())
+            ->setFileMimeType(self::getFile()->getMimeType())
+            ->setFileBaseName(self::getFile()->getBaseName())
+            ->setFileImporterArguments(self::getFile()->getImporterArguments())
+            ->setBrand(\Ivoz\Provider\Domain\Model\Brand\Brand::entityToDto(self::getBrand(), $depth));
+    }
+
+    /**
+     * @return array
+     */
+    protected function __toArray()
+    {
+        return [
+            'status' => self::getStatus(),
+            'nameEn' => self::getName()->getEn(),
+            'nameEs' => self::getName()->getEs(),
+            'descriptionEn' => self::getDescription()->getEn(),
+            'descriptionEs' => self::getDescription()->getEs(),
+            'fileFileSize' => self::getFile()->getFileSize(),
+            'fileMimeType' => self::getFile()->getMimeType(),
+            'fileBaseName' => self::getFile()->getBaseName(),
+            'fileImporterArguments' => self::getFile()->getImporterArguments(),
+            'brandId' => self::getBrand() ? self::getBrand()->getId() : null
+        ];
+    }
+
+
+    // @codeCoverageIgnoreStart
+
+    /**
+     * Set status
+     *
+     * @param string $status
+     *
+     * @return self
+     */
+    public function setStatus($status = null)
+    {
+        if (!is_null($status)) {
+            Assertion::maxLength($status, 20, 'status value "%s" is too long, it should have no more than %d characters, but has %d characters.');
+        Assertion::choice($status, array (
+          0 => 'waiting',
+          1 => 'inProgress',
+          2 => 'imported',
+          3 => 'error',
+        ), 'statusvalue "%s" is not an element of the valid values: %s');
+        }
+
+        $this->status = $status;
+
+        return $this;
+    }
+
+    /**
+     * Get status
+     *
+     * @return string
+     */
+    public function getStatus()
+    {
+        return $this->status;
+    }
+
+    /**
+     * Set brand
+     *
+     * @param \Ivoz\Provider\Domain\Model\Brand\BrandInterface $brand
+     *
+     * @return self
+     */
+    public function setBrand(\Ivoz\Provider\Domain\Model\Brand\BrandInterface $brand)
+    {
+        $this->brand = $brand;
+
+        return $this;
+    }
+
+    /**
+     * Get brand
+     *
+     * @return \Ivoz\Provider\Domain\Model\Brand\BrandInterface
+     */
+    public function getBrand()
+    {
+        return $this->brand;
+    }
+
+    /**
+     * Set name
+     *
+     * @param \Ivoz\Provider\Domain\Model\DestinationRateGroup\Name $name
+     *
+     * @return self
+     */
+    public function setName(Name $name)
+    {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    /**
+     * Get name
+     *
+     * @return \Ivoz\Provider\Domain\Model\DestinationRateGroup\Name
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+
+    /**
+     * Set description
+     *
+     * @param \Ivoz\Provider\Domain\Model\DestinationRateGroup\Description $description
+     *
+     * @return self
+     */
+    public function setDescription(Description $description)
+    {
+        $this->description = $description;
+
+        return $this;
+    }
+
+    /**
+     * Get description
+     *
+     * @return \Ivoz\Provider\Domain\Model\DestinationRateGroup\Description
+     */
+    public function getDescription()
+    {
+        return $this->description;
+    }
+
+    /**
+     * Set file
+     *
+     * @param \Ivoz\Provider\Domain\Model\DestinationRateGroup\File $file
+     *
+     * @return self
+     */
+    public function setFile(File $file)
+    {
+        $this->file = $file;
+
+        return $this;
+    }
+
+    /**
+     * Get file
+     *
+     * @return \Ivoz\Provider\Domain\Model\DestinationRateGroup\File
+     */
+    public function getFile()
+    {
+        return $this->file;
+    }
+
+    // @codeCoverageIgnoreEnd
+}
+
