+++ library/Ivoz/Provider/Domain/Model/DestinationRateGroup/DestinationRateGroupDtoAbstract.php
@@ -1,0 +1,426 @@
+<?php
+
+namespace Ivoz\Provider\Domain\Model\DestinationRateGroup;
+
+use Ivoz\Core\Application\DataTransferObjectInterface;
+use Ivoz\Core\Application\ForeignKeyTransformerInterface;
+use Ivoz\Core\Application\CollectionTransformerInterface;
+use Ivoz\Core\Application\Model\DtoNormalizer;
+
+/**
+ * @codeCoverageIgnore
+ */
+abstract class DestinationRateGroupDtoAbstract implements DataTransferObjectInterface
+{
+    /**
+     * @var string
+     */
+    private $status;
+
+    /**
+     * @var integer
+     */
+    private $id;
+
+    /**
+     * @var string
+     */
+    private $nameEn;
+
+    /**
+     * @var string
+     */
+    private $nameEs;
+
+    /**
+     * @var string
+     */
+    private $descriptionEn;
+
+    /**
+     * @var string
+     */
+    private $descriptionEs;
+
+    /**
+     * @var integer
+     */
+    private $fileFileSize;
+
+    /**
+     * @var string
+     */
+    private $fileMimeType;
+
+    /**
+     * @var string
+     */
+    private $fileBaseName;
+
+    /**
+     * @var array
+     */
+    private $fileImporterArguments;
+
+    /**
+     * @var \Ivoz\Provider\Domain\Model\Brand\BrandDto | null
+     */
+    private $brand;
+
+    /**
+     * @var \Ivoz\Provider\Domain\Model\DestinationRate\DestinationRateDto[] | null
+     */
+    private $destinationRates = null;
+
+
+    use DtoNormalizer;
+
+    public function __construct($id = null)
+    {
+        $this->setId($id);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public static function getPropertyMap(string $context = '')
+    {
+        if ($context === self::CONTEXT_COLLECTION) {
+            return ['id' => 'id'];
+        }
+
+        return [
+            'status' => 'status',
+            'id' => 'id',
+            'name' => ['en','es'],
+            'description' => ['en','es'],
+            'file' => ['fileSize','mimeType','baseName','importerArguments'],
+            'brandId' => 'brand'
+        ];
+    }
+
+    /**
+     * @return array
+     */
+    public function toArray($hideSensitiveData = false)
+    {
+        return [
+            'status' => $this->getStatus(),
+            'id' => $this->getId(),
+            'name' => [
+                'en' => $this->getNameEn(),
+                'es' => $this->getNameEs()
+            ],
+            'description' => [
+                'en' => $this->getDescriptionEn(),
+                'es' => $this->getDescriptionEs()
+            ],
+            'file' => [
+                'fileSize' => $this->getFileFileSize(),
+                'mimeType' => $this->getFileMimeType(),
+                'baseName' => $this->getFileBaseName(),
+                'importerArguments' => $this->getFileImporterArguments()
+            ],
+            'brand' => $this->getBrand(),
+            'destinationRates' => $this->getDestinationRates()
+        ];
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function transformForeignKeys(ForeignKeyTransformerInterface $transformer)
+    {
+        $this->brand = $transformer->transform('Ivoz\\Provider\\Domain\\Model\\Brand\\Brand', $this->getBrandId());
+        if (!is_null($this->destinationRates)) {
+            $items = $this->getDestinationRates();
+            $this->destinationRates = [];
+            foreach ($items as $item) {
+                $this->destinationRates[] = $transformer->transform(
+                    'Ivoz\\Provider\\Domain\\Model\\DestinationRate\\DestinationRate',
+                    $item->getId() ?? $item
+                );
+            }
+        }
+
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function transformCollections(CollectionTransformerInterface $transformer)
+    {
+        $this->destinationRates = $transformer->transform(
+            'Ivoz\\Provider\\Domain\\Model\\DestinationRate\\DestinationRate',
+            $this->destinationRates
+        );
+    }
+
+    /**
+     * @param string $status
+     *
+     * @return static
+     */
+    public function setStatus($status = null)
+    {
+        $this->status = $status;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getStatus()
+    {
+        return $this->status;
+    }
+
+    /**
+     * @param integer $id
+     *
+     * @return static
+     */
+    public function setId($id = null)
+    {
+        $this->id = $id;
+
+        return $this;
+    }
+
+    /**
+     * @return integer
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param string $nameEn
+     *
+     * @return static
+     */
+    public function setNameEn($nameEn = null)
+    {
+        $this->nameEn = $nameEn;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getNameEn()
+    {
+        return $this->nameEn;
+    }
+
+    /**
+     * @param string $nameEs
+     *
+     * @return static
+     */
+    public function setNameEs($nameEs = null)
+    {
+        $this->nameEs = $nameEs;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getNameEs()
+    {
+        return $this->nameEs;
+    }
+
+    /**
+     * @param string $descriptionEn
+     *
+     * @return static
+     */
+    public function setDescriptionEn($descriptionEn = null)
+    {
+        $this->descriptionEn = $descriptionEn;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getDescriptionEn()
+    {
+        return $this->descriptionEn;
+    }
+
+    /**
+     * @param string $descriptionEs
+     *
+     * @return static
+     */
+    public function setDescriptionEs($descriptionEs = null)
+    {
+        $this->descriptionEs = $descriptionEs;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getDescriptionEs()
+    {
+        return $this->descriptionEs;
+    }
+
+    /**
+     * @param integer $fileFileSize
+     *
+     * @return static
+     */
+    public function setFileFileSize($fileFileSize = null)
+    {
+        $this->fileFileSize = $fileFileSize;
+
+        return $this;
+    }
+
+    /**
+     * @return integer
+     */
+    public function getFileFileSize()
+    {
+        return $this->fileFileSize;
+    }
+
+    /**
+     * @param string $fileMimeType
+     *
+     * @return static
+     */
+    public function setFileMimeType($fileMimeType = null)
+    {
+        $this->fileMimeType = $fileMimeType;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getFileMimeType()
+    {
+        return $this->fileMimeType;
+    }
+
+    /**
+     * @param string $fileBaseName
+     *
+     * @return static
+     */
+    public function setFileBaseName($fileBaseName = null)
+    {
+        $this->fileBaseName = $fileBaseName;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getFileBaseName()
+    {
+        return $this->fileBaseName;
+    }
+
+    /**
+     * @param array $fileImporterArguments
+     *
+     * @return static
+     */
+    public function setFileImporterArguments($fileImporterArguments = null)
+    {
+        $this->fileImporterArguments = $fileImporterArguments;
+
+        return $this;
+    }
+
+    /**
+     * @return array
+     */
+    public function getFileImporterArguments()
+    {
+        return $this->fileImporterArguments;
+    }
+
+    /**
+     * @param \Ivoz\Provider\Domain\Model\Brand\BrandDto $brand
+     *
+     * @return static
+     */
+    public function setBrand(\Ivoz\Provider\Domain\Model\Brand\BrandDto $brand = null)
+    {
+        $this->brand = $brand;
+
+        return $this;
+    }
+
+    /**
+     * @return \Ivoz\Provider\Domain\Model\Brand\BrandDto
+     */
+    public function getBrand()
+    {
+        return $this->brand;
+    }
+
+    /**
+     * @param integer $id | null
+     *
+     * @return static
+     */
+    public function setBrandId($id)
+    {
+        $value = !is_null($id)
+            ? new \Ivoz\Provider\Domain\Model\Brand\BrandDto($id)
+            : null;
+
+        return $this->setBrand($value);
+    }
+
+    /**
+     * @return integer | null
+     */
+    public function getBrandId()
+    {
+        if ($dto = $this->getBrand()) {
+            return $dto->getId();
+        }
+
+        return null;
+    }
+
+    /**
+     * @param array $destinationRates
+     *
+     * @return static
+     */
+    public function setDestinationRates($destinationRates = null)
+    {
+        $this->destinationRates = $destinationRates;
+
+        return $this;
+    }
+
+    /**
+     * @return array
+     */
+    public function getDestinationRates()
+    {
+        return $this->destinationRates;
+    }
+}
+
+
