+++ library/Ivoz/Provider/Domain/Model/DestinationRateGroup/DestinationRateGroupRepository.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Ivoz\Provider\Domain\Model\DestinationRateGroup;
+
+use Doctrine\Common\Persistence\ObjectRepository;
+use Doctrine\Common\Collections\Selectable;
+
+interface DestinationRateGroupRepository extends  ObjectRepository, Selectable
+{
+
+}
+
+
