+++ library/Ivoz/Provider/Domain/Model/DestinationRateGroup/Name.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace Ivoz\Provider\Domain\Model\DestinationRate;
+namespace Ivoz\Provider\Domain\Model\DestinationRateGroup;
 
 use Assert\Assertion;
 use Ivoz\Core\Domain\Model\EntityInterface;
 
 /**
- * Description
+ * Name
  * @codeCoverageIgnore
  */
-class Description
+class Name
 {
     /**
-     * column: description_en
+     * column: name_en
      * @var string
      */
     protected $en;
 
     /**
-     * column: description_es
+     * column: name_es
      * @var string
      */
     protected $es;
@@ -45,7 +45,7 @@ class Description
     protected function setEn($en)
     {
         Assertion::notNull($en, 'en value "%s" is null, but non null value was expected.');
-        Assertion::maxLength($en, 255, 'en value "%s" is too long, it should have no more than %d characters, but has %d characters.');
+        Assertion::maxLength($en, 55, 'en value "%s" is too long, it should have no more than %d characters, but has %d characters.');
 
         $this->en = $en;
 
@@ -72,7 +72,7 @@ class Description
     protected function setEs($es)
     {
         Assertion::notNull($es, 'es value "%s" is null, but non null value was expected.');
-        Assertion::maxLength($es, 255, 'es value "%s" is too long, it should have no more than %d characters, but has %d characters.');
+        Assertion::maxLength($es, 55, 'es value "%s" is too long, it should have no more than %d characters, but has %d characters.');
 
         $this->es = $es;
