+++ library/Ivoz/Provider/Infrastructure/Persistence/Doctrine/DestinationRateGroupDoctrineRepository.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Ivoz\Provider\Infrastructure\Persistence\Doctrine;
+
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Ivoz\Provider\Domain\Model\DestinationRateGroup\DestinationRateGroupRepository;
+use Ivoz\Provider\Domain\Model\DestinationRateGroup\DestinationRateGroup;
+use Symfony\Bridge\Doctrine\RegistryInterface;
+
+/**
+ * DestinationRateGroupDoctrineRepository
+ *
+ * This class was generated by the Doctrine ORM. Add your own custom
+ * repository methods below.
+ */
+class DestinationRateGroupDoctrineRepository extends ServiceEntityRepository implements DestinationRateGroupRepository
+{
+    public function __construct(RegistryInterface $registry)
+    {
+        parent::__construct($registry, DestinationRateGroup::class);
+    }
+}
+
